split out ncpptest from test_main
