+++ test/src/realtime/RtClientTest.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "test_main.h"
+#include "NCppTest.h"
 #include "nakama-cpp/realtime/NRtDefaultClientListener.h"
 
 namespace Nakama {
