+++ test/src/test_authentication.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "globals.h"
-#include "test_main.h"
+#include "NCppTest.h"
 
 namespace Nakama {
 namespace Test {
